requirements.txt
Flask==3.1.1
Flask-SQLAlchemy==3.1.1
Flask-Migrate==4.0.5
Flask-Login==0.6.3
psycopg2-binary==2.9.10
python-dotenv==1.0.0
Werkzeug==3.1.1
pyrogram==2.0.106
tgcrypto==1.2.5
apscheduler==3.10.4
requests==2.31.0
click==8.1.7
gunicorn==21.2.0
python-telegram-bot==20.7